+++ web/src/components/BusinessCard/BusinessCard.test.tsx
@@ -1,14 +1,15 @@
 import { render } from '@redwoodjs/testing/web'
 
-import WorkPlacesCard from './WorkPlacesCard'
+import WorkPlacesCard from './BusinessCard'
+import BusinessCard from './BusinessCard'
 
 //   Improve this test with help from the Redwood Testing Doc:
 //    https://redwoodjs.com/docs/testing#testing-components
 
-describe('WorkPlacesCard', () => {
+describe('BusinessCard', () => {
   it('renders successfully', () => {
     expect(() => {
-      render(<WorkPlacesCard />)
+      render(<BusinessCard />)
     }).not.toThrow()
   })
 })
